requirements.txt
customtkinter>=5.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
ocrmypdf>=17.0.0
